refactor: 출력형식 수정
